Move input to console package
